pyproject.toml
[tool.poetry]
name = "generalqbot"
version = "0.0.1"
description = "一个通用QQ机器人"
authors = ["XYCode-Kerman <xycode-xyc@outlook.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<=3.10"
EdgeGPT = "^0.3.6"
yiri-mirai = "^0.2.7"
yiri-mirai-trigger = "^0.3.1"
typing = "3.7.4.3"
typing_extensions = "4.6.1"
tensorflow = "^2.4.1"
tencent-cloud-sdk = "^0.2.7"
pymongo = "^4.3.3"
PyJWT = "^2.6.0"
python-dotenv = "^1.0.0"
httpx = "0.24.1"
click = "8.1.3"
socks = "0"
jieba = "^0.42.1"
tencentcloud-sdk-python-tms = "3.0.885"
tencentcloud-sdk-python-common = "3.0.885"
openai = "^0.27.7"
apscheduler = "^3.10.1"
recognizers-text = "^1.0.2a2"
recognizers-number-with-unit = "^1.0.0a1"
emoji = "1.1.0"
objprint = "^0.2.2"
socksio = "^1.0.0"
flask = {extras = ["async"], version = "^2.3.2"}
wordcloud = "^1.9.2"
nltk = "3.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
anyio = "^3.7.0"
pytest-asyncio = "^0.21.0"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.0"
twisted = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
